# src/main/resources/swagger/project1.yml
openapi: 3.0.3
info:
  title: Servicio de Gestión Bancaria
  description: El proyecto gestiona información de clientes, productos financieros, y transacciones, abarcando cuentas bancarias, créditos y tarjetas.
  termsOfService: http://swagger.io/terms/
  contact:
    email: anacarenasr@sgmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: http://localhost:8080/api/v1
    description: Servidor de desarrollo
...
paths:
...
  /credit:
    get:
      tags:
        - credits
      summary: Obtener lista de créditos
      description: Retorna una lista de todos los créditos registrados.
      responses:
        '200':
          description: Lista de créditos obtenida exitosamente
          content:
            application/json:
              example:
                - id: 1
                  customerId: 1
                  typeAccount: credit
                  descriptTypeAccount: Crédito Personal
                  creditAmount: 10000.00
                  existingAmount: 5000.00
                  creditDate: 2024-12-31
                  typeCustomer: individual
                  interestRate: 0.1
                  paymentFrequency: monthly
                  installmentCount: 12
                  remainingInstallments: 6
                  nextPaymentDate: 2024-07-31
                  status: active
                - id: 2
                  customerId: 2
                  typeAccount: credit
                  descriptTypeAccount: Crédito Empresarial
                  creditAmount: 50000.00
                  existingAmount: 20000.00
                  creditDate: 2024-12-31
                  typeCustomer: business
                  interestRate: 0.08
                  paymentFrequency: quarterly
                  installmentCount: 24
                  remainingInstallments: 12
                  nextPaymentDate: 2024-10-31
                  status: active

    post:
      tags:
        - credits
      summary: Crear un nuevo crédito
      description: Permite registrar un nuevo crédito en el sistema.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credit'
      responses:
        '201':
          description: Crédito creado exitosamente
        '400':
          description: Solicitud incorrecta
          content:
            application/json:
              example:
                message: Error de validación en la solicitud
  /credit/{id}:
    get:
      tags:
        - credits
      summary: Obtener detalles de un crédito por ID
      description: Retorna los detalles de un crédito específico según su ID.
      parameters:
        - name: id
          in: path
          description: ID único del crédito
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Detalles del crédito obtenidos exitosamente
          content:
            application/json:
              example:
                id: 1
                customerId: 1
                typeAccount: credit
                descriptTypeAccount: Crédito Personal
                creditAmount: 10000.00
                existingAmount: 5000.00
                creditDate: 2024-12-31
                typeCustomer: individual
                interestRate: 0.1
                paymentFrequency: monthly
                installmentCount: 12
                remainingInstallments: 6
                nextPaymentDate: 2024-07-31
                status: active

    put:
      tags:
        - credits
      summary: Actualizar un crédito existente
      description: Permite actualizar la información de un crédito existente en el sistema.
      parameters:
        - name: id
          in: path
          description: ID único del crédito
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credit'
      responses:
        '200':
          description: Crédito actualizado exitosamente
        '400':
          description: Solicitud incorrecta
          content:
            application/json:
              example:
                message: Error de validación en la solicitud

    delete:
      tags:
        - credits
      summary: Eliminar un crédito
      description: Permite eliminar un crédito existente del sistema.
      parameters:
        - name: id
          in: path
          description: ID único del crédito
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Crédito eliminado exitosamente
...
components:
  schemas:
...
    Credit:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID único del crédito.
        customerId:
          type: integer
          description: ID del cliente asociado.
        typeAccount:
          type: string
          description: Tipo de cuenta asociada.
        descriptTypeAccount:
          type: string
          description: Descripción del tipo de cuenta.
        creditAmount:
          type: number
          format: double
          description: Monto total del crédito.
        existingAmount:
          type: number
          format: double
          description: Monto actual del crédito.
        creditDate:
          type: string
          format: date
          description: Fecha de vencimiento del crédito.
        typeCustomer:
          type: string
          enum: [individual, business]
          description: Tipo de cliente (individual o empresarial).
        interestRate:
          type: number
          format: double
          description: Tasa de interés aplicada al crédito.
        paymentFrequency:
          type: string
          enum: [monthly, quarterly, annually]
          description: Frecuencia de pago (mensual, trimestral, anual).
        installmentCount:
          type: integer
          description: Número total de cuotas del crédito.
        remainingInstallments:
          type: integer
          description: Número de cuotas restantes.
        nextPaymentDate:
          type: string
          format: date
          description: Próxima fecha de pago.
        status:
          type: string
          enum: [active, closed, overdue]
          description: Estado del crédito (activo, cerrado, vencido).

      required:
        - customerId
        - typeAccount
        - descriptTypeAccount
        - creditAmount
        - existingAmount
        - creditDate
        - typeCustomer
        - interestRate
        - paymentFrequency
        - installmentCount
        - remainingInstallments
        - nextPaymentDate
        - status
...